+++ CHANGES.md
@@ -12,6 +12,7 @@
 * Lib: add PerformanceObserver API (#1164)
 * Lib: add CSSStyleDeclaration.{setProperty, getPropertyValue, getPropertyPriority, removeProperty} (#1170)
 * PPX: json can now be derived for mutable records (#1184)
+* Runtime: use crypto.getRandomValues when available (#1209)
 
 ## Bug fixes
 * Compiler: fix sourcemap warning for empty cma (#1169)